requirements.txt
streamlit
langchain
pypdf2
faiss-cpu
openai
python-dotenv
